refactor: extract shared lyrics download service
