+++ src/ui/services/__init__.py
@@ -1,0 +1,1 @@
+"""UI services package."""
